+++ src/Public/Get-ShlinkDomains.ps1
@@ -1,19 +1,19 @@
-function Get-ShlinkTags {
+function Get-ShlinkDomains {
     <#
     .SYNOPSIS
-        Returns the list of all tags used in any short URL, including stats and ordered by name.
+        Returns the list of all domains ever used, with a flag that tells if they are the default domain
     .DESCRIPTION
-        Returns the list of all tags used in any short URL, including stats and ordered by name.
+        Returns the list of all domains ever used, with a flag that tells if they are the default domain
     .PARAMETER ShlinkServer
         The URL of your Shlink server (including schema). For example "https://example.com".
         It is not required to use this parameter for every use of this function. When it is used once for any of the functions in the PSShlink module, its value is retained throughout the life of the PowerShell session and its value is only accessible within the module's scope.
     .PARAMETER ShlinkApiKey
         A SecureString object of your Shlink server's API key.
-        It is not required to use this parameter for every use of this function. When it is used once for any of the functions in the PSShlink module, its value is retained throughout the life of the PowerShell session and its value is only accessible within the module's scope.
+        It is not required to use this parameter for every use of this function. When it is used once for any of the functions in the PSShlink module, its value is retained throughout the life of the PowerShell session and its value is only accessible within the module's scope.    
     .EXAMPLE
-        PS C:\> Get-ShlinkTags
-        
-        Returns the list of all tags used in any short URL, including stats and ordered by name.
+        PS C:\> Get-ShlinkDomains
+
+        Returns the list of all domains ever used, with a flag that tells if they are the default domain
     .INPUTS
         This function does not accept pipeline input.
     .OUTPUTS
@@ -34,17 +34,13 @@ function Get-ShlinkTags {
     catch {
         Write-Error -ErrorRecord $_ -ErrorAction "Stop"
     }
-    
-    $QueryString = [System.Web.HttpUtility]::ParseQueryString('')
-
-    $QueryString.Add("withStats", "true")
 
     $Params = @{
-        Endpoint     = "tags"
-        PropertyTree = "tags", "stats"
+        Endpoint     = "domains"
+        PropertyTree = "domains"
+        PSTypeName   = "PSShlinkDomains"
+        ErrorAction  = "Stop"
     }
-
-    $Params["Query"] = $QueryString
 
     try {
         InvokeShlinkRestMethod @Params