+++ tests/Public/New-ShlinkUrl.Tests.ps1
@@ -1,7 +1,9 @@
 BeforeAll {
-    Import-Module $PSScriptRoot/../../../build/PSShlink/PSShlink.psd1 -Force
-    $env:ShlinkServer = "https://acook.io"
-    $env:ShlinkAPIKey = (Get-Secure "Shlink").Password | ConvertFrom-SecureString
+    # Check if module is already imported, as it can be via VSCode task where you can choose what code base to test
+    # and you might not want to cloober it with the non-built code
+    if (-not (Get-Module PSShlink)) {
+        Import-Module $PSScriptRoot/../../src/PSShlink.psd1 -Force
+    }
 }
 
 Describe "Create new short URLs" {
@@ -34,7 +36,7 @@ Describe "Create new short URLs" {
 
     It "New invalid short URL with validation off" {
         $Guid = (New-Guid).Guid
-        # Multiply guid string by 3 because ideally we really do not want a valid URL
+        # Multiply guid string by 3 because ideally we *really* do not want a valid URL
         $Url = 'https://{0}.com' -f ($Guid * 3)
 
         { Invoke-WebRequest -Uri $Url -ErrorAction 'Stop' } | Should -Throw -ExceptionType ([System.Net.Http.HttpRequestException])